优化gic
