+++ crates/sparreal-kernel/src/mem/mmu/mod.rs
@@ -167,6 +167,8 @@ pub fn set_user_table(addr: usize) {
 pub fn get_user_table() -> usize {
     MMUImpl::get_user_table()
 }
+
+#[allow(unused)]
 pub(crate) fn flush_tlb(addr: *const u8) {
     unsafe { MMUImpl::flush_tlb(addr) };
 }
