+++ crates/sparreal-rt/src/arch/aarch64/gic/gic_v2.rs
@@ -1,53 +1,58 @@
+use alloc::{boxed::Box, format, string::ToString, sync::Arc, vec::Vec};
 use core::{cell::UnsafeCell, ptr::NonNull};
 
-use alloc::{boxed::Box, format, string::ToString, sync::Arc, vec::Vec};
-use arm_gic_driver::{GicGeneric, GicV3, IntId, Trigger};
+use arm_gic_driver::{GicGeneric, Trigger};
 use sparreal_kernel::{
     driver_interface::{
-        self, DriverGeneric, DruverResult, ProbeFn, RegAddress,
+        self, DriverGeneric, DriverResult, ProbeFn, RegAddress,
         interrupt_controller::{self, CpuId, InterruptControllerPerCpu},
     },
     mem::iomap,
 };
 use sparreal_macros::module_driver;
 
+use super::*;
+
 module_driver!(
-    name: "GICv3",
-    compatibles: "arm,gic-v3\n",
-    probe: ProbeFn::InterruptController(probe_gic),
+    name: "GICv2",
+    compatibles: "arm,cortex-a15-gic\n",
+    probe: ProbeFn::InterruptController(probe_gic_v2),
 );
 
-struct Gic {
+struct GicV2 {
     gicd: NonNull<u8>,
-    gicr: NonNull<u8>,
-    gic: Arc<UnsafeCell<Option<GicV3>>>,
+    gicc: NonNull<u8>,
+    gic: Arc<UnsafeCell<Option<arm_gic_driver::GicV2>>>,
 }
 
-unsafe impl Send for Gic {}
+unsafe impl Send for GicV2 {}
 
-impl Gic {
-    fn new(gicd: NonNull<u8>, gicr: NonNull<u8>) -> Self {
+impl GicV2 {
+    #[allow(clippy::arc_with_non_send_sync)]
+    fn new(gicd: NonNull<u8>, gicc: NonNull<u8>) -> Self {
         Self {
             gic: Arc::new(UnsafeCell::new(None)),
             gicd,
-            gicr,
+            gicc,
         }
     }
 }
 
-struct GicPerCpu(Arc<UnsafeCell<Option<GicV3>>>);
+struct GicV2PerCpu(Arc<UnsafeCell<Option<arm_gic_driver::GicV2>>>);
 
-unsafe impl Send for GicPerCpu {}
+unsafe impl Send for GicV2PerCpu {}
 
-impl GicPerCpu {
-    fn get_mut(&self) -> &mut GicV3 {
+impl GicV2PerCpu {
+    fn get_mut(&self) -> &mut arm_gic_driver::GicV2 {
         unsafe { &mut *self.0.get() }.as_mut().unwrap()
     }
 }
 
-impl InterruptControllerPerCpu for GicPerCpu {
+impl InterruptControllerPerCpu for GicV2PerCpu {
     fn get_and_acknowledge_interrupt(&self) -> Option<interrupt_controller::IrqId> {
-        self.get_mut()
+        unsafe { &mut *self.0.get() }
+            .as_mut()
+            .unwrap()
             .get_and_acknowledge_interrupt()
             .map(|id| (id.to_u32() as usize).into())
     }
@@ -90,51 +95,45 @@ impl InterruptControllerPerCpu for GicPerCpu {
             .collect::<Vec<_>>();
 
         self.get_mut().set_bind_cpu(convert_id(irq), &id_list);
-
-        todo!()
     }
 }
 
-fn convert_id(irq: interrupt_controller::IrqId) -> IntId {
-    let id: usize = irq.into();
-    unsafe { IntId::raw(id as u32) }
-}
-
-impl DriverGeneric for Gic {
+impl DriverGeneric for GicV2 {
     fn name(&self) -> alloc::string::String {
-        "GICv3".to_string()
+        "GICv2".to_string()
     }
 
     fn open(&mut self) -> Result<(), alloc::string::String> {
-        let gic = GicV3::new(self.gicd, self.gicr).map_err(|e| format!("{:?}", e))?;
+        let gic =
+            arm_gic_driver::GicV2::new(self.gicd, self.gicc).map_err(|e| format!("{:?}", e))?;
         unsafe { &mut *self.gic.get() }.replace(gic);
 
         Ok(())
     }
 }
 
-impl interrupt_controller::InterruptController for Gic {
+impl interrupt_controller::InterruptController for GicV2 {
     fn current_cpu_setup(&self) -> Box<dyn interrupt_controller::InterruptControllerPerCpu> {
         unsafe { &mut *self.gic.get() }
             .as_mut()
             .unwrap()
             .current_cpu_setup();
-        Box::new(GicPerCpu(self.gic.clone()))
+        Box::new(GicV2PerCpu(self.gic.clone()))
     }
 
     fn parse_fdt_config(
         &self,
         prop_interupt: &[usize],
-    ) -> DruverResult<driver_interface::IrqConfig> {
-        todo!()
+    ) -> DriverResult<driver_interface::IrqConfig> {
+        Ok(fdt_itr_to_config(&prop_interupt))
     }
 }
 
-fn probe_gic(regs: Vec<RegAddress>) -> interrupt_controller::BoxedDriver {
+fn probe_gic_v2(regs: Vec<RegAddress>) -> interrupt_controller::BoxedDriver {
     let gicd_reg = regs[0];
     let gicc_reg = regs[1];
     let gicd = iomap(gicd_reg.addr.into(), gicd_reg.size.unwrap_or(0x1000));
-    let gicr = iomap(gicc_reg.addr.into(), gicc_reg.size.unwrap_or(0x1000));
+    let gicc = iomap(gicc_reg.addr.into(), gicc_reg.size.unwrap_or(0x1000));
 
-    Box::new(Gic::new(gicd, gicr))
+    Box::new(GicV2::new(gicd, gicc))
 }
